Revert "FISH-6033 renamed to fix deploy"
This reverts commit ffeddbc8fd0b72d96b20f49df0f92d17c5b1d1a1.
